addressing comments. adding to CI
