+++ aptos-move/aptos-gas-calibration/src/measurements_helpers.rs
@@ -5,7 +5,10 @@ use crate::measurements::GasMeasurements;
 use anyhow::{anyhow, Result};
 use aptos_cached_packages::aptos_stdlib;
 use aptos_framework::BuiltPackage;
-use aptos_language_e2e_tests::{account::Account, executor::FakeExecutor};
+use aptos_language_e2e_tests::{
+    account::Account,
+    executor::{ExecFuncTimerDynamicArgs, FakeExecutor, GasMeterType},
+};
 use aptos_types::transaction::TransactionPayload;
 use move_binary_format::{file_format::SignatureToken, CompiledModule};
 use move_core_types::{
@@ -97,6 +100,8 @@ pub fn execute_user_txn(
         vec![],
         args,
         iterations,
+        ExecFuncTimerDynamicArgs::NoArgs,
+        GasMeterType::UnmeteredGasMeter,
     );
     println!("running time (microseconds): {}", elapsed);
     elapsed
